requirement.txt
aanalytics2==0.4.4
certifi==2025.6.15
charset-normalizer==3.4.2
ckanapi==4.8
contourpy==1.3.2
cycler==0.12.1
docopt==0.6.2
fonttools==4.58.4
idna==3.10
kaleido==1.2.0
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pillow==12.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
text-unidecode==1.3
tzdata==2025.2
urllib3==2.5.0
plotly==6.5.0
